requirements.txt
asgiref==3.3.1
cycler==0.10.0
dj-database-url==0.5.0
Django==3.1.4
django-heroku==0.3.1
django-pandas==0.6.2
et-xmlfile==1.0.1
gunicorn==20.0.4
jdcal==1.4.1
kiwisolver==1.3.1
matplotlib==3.3.3
japanize-matplotlib==1.1.3
mysql-connector-python==8.0.22
mysqlclient==2.0.2
numpy==1.19.4
openpyxl==3.0.5
pandas==0.24.2
Pillow==8.0.1
protobuf==3.14.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
whitenoise==5.2.0
psycopg2-binary==2.8.6
xlsxwriter==1.3.7
requests
line-bot-sdk
scikit-learn
xlrd==1.2.0
selenium
lxml
boto3
django-storages
devtools
jpholiday
python-dotenv
msal==1.10.0
